+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- version of the app is in ElectronMainApp/package.json -->
 
 
+## [Unreleased]
+
+### Changed
+
+- Updated [`@adguard/scriptlets`] to v1.9.96
+
+
 ## [1.11.16] - 2023-11-21
 
 ### Changed
@@ -55,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated ExtendedCss to v2.0.52
 
 
+[Unreleased]: https://github.com/AdguardTeam/AdGuardForSafari/compare/v1.11.16.324-release...HEAD
 [1.11.16]: https://github.com/AdguardTeam/AdGuardForSafari/compare/v1.11.15.309-release...v1.11.16.324-release
 [1.11.15]: https://github.com/AdguardTeam/AdGuardForSafari/compare/v1.11.14.301-release...v1.11.15.309-release
 [1.11.14]: https://github.com/AdguardTeam/AdGuardForSafari/compare/v1.11.13.297-release...v1.11.14.301-release